// silap-colmirwcl/src/app/modules/silapapi/model/apiMessage.ts
/**
 * 
 * 
 *
 * 
 * 
 *
 * NOTE: This class is auto generated by the swagger code generator program.
 * https://github.com/swagger-api/swagger-codegen.git
 * Do not edit the class manually.
 */

export interface ApiMessage { 
    code?: string;
    error?: boolean;
    message?: string;
    tipo?: string;
    title?: string;
}